requirements.txt
llama-cpp-python
gradio
requests
python-dotenv
